refactor(activerecord): move JoinDependency, ConnectionPool, DatabaseConfigurations to correct paths
Move 3 classes back to their Rails-canonical parent file paths (they
were incorrectly placed in subdirectories in earlier PRs due to an
api:compare class-reopening bug):

- associations/join-dependency/join-association.ts → associations/join-dependency.ts
- connection-adapters/abstract/connection-pool/queue.ts → connection-adapters/abstract/connection-pool.ts
- database-configurations/connection-url-resolver.ts → database-configurations.ts

The remaining 1 misplaced (Cipher) is a Ruby autoloading artifact — Rails
defines the Cipher namespace in encryption.rb, not a pattern that maps
to TS modules.

api:compare: 61/597 (10.2%), 1 misplaced (was 4).
