+++ packages/activerecord/src/connection-pool.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { ConnectionPool } from "./connection-adapters/abstract/connection-pool/queue.js";
+import { ConnectionPool } from "./connection-adapters/abstract/connection-pool.js";
 import { HashConfig } from "./database-configurations/hash-config.js";
 import { createTestAdapter } from "./test-adapter.js";
