+++ packages/activerecord/src/index.ts
@@ -96,8 +96,8 @@ export { DatabaseConfig } from "./database-configurations/database-config.js";
 export type { DatabaseConfigOptions } from "./database-configurations/database-config.js";
 export { HashConfig } from "./database-configurations/hash-config.js";
 export { UrlConfig } from "./database-configurations/url-config.js";
-export { DatabaseConfigurations } from "./database-configurations/connection-url-resolver.js";
-export { ConnectionPool } from "./connection-adapters/abstract/connection-pool/queue.js";
+export { DatabaseConfigurations } from "./database-configurations.js";
+export { ConnectionPool } from "./connection-adapters/abstract/connection-pool.js";
 export { ConnectionHandler } from "./connection-adapters/abstract/connection-handler.js";
 export { DatabaseTasks } from "./tasks/database-tasks.js";
 export type { DatabaseTaskHandler } from "./tasks/database-tasks.js";
